pyproject.toml
[tool.poetry]
name = "backend_ht5"
version = "0.1.0"
description = ""
authors = ["Evhen Hordiienko <hoodback@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.3"
Django = "3.1.5"
celery = "5.0.5"
requests = "^2.25.1"
django-extensions = "^3.1.0"
gunicorn = "^20.0.4"
django-redis = "^4.12.1"
django-cors-headers = "^3.7.0"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"
django-filter = "^2.4.0"
freezegun = "^1.1.0"
coverage = "^5.5"

[tool.poetry.dev-dependencies]
flake8 = "3.8.4"
Faker = "5.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
